put chat logic in a separate file so it could be reused in multiple places added retry button on error banner
