+++ client/src/components/chat/chat.css
@@ -1,4 +1,4 @@
-/* ── LOCAL TOKENS ───────────────────────────────────────────────── */
+/* LOCAL TOKENS */
 :root {
     --chat-accent:        #00e5ff;
     --chat-accent-bg:     rgba(0, 229, 255, 0.07);
@@ -12,7 +12,7 @@
     --chat-sidebar-w:     220px;
 }
 
-/* ── LAYOUT ─────────────────────────────────────────────────────── */
+/* LAYOUT */
 .chat-container {
     display: grid;
     grid-template-columns: var(--chat-sidebar-w) 1fr;
@@ -23,7 +23,7 @@
     height: var(--chat-height);
 }
 
-/* ── SIDEBAR ────────────────────────────────────────────────────── */
+/* SIDEBAR */
 .chat-sidebar {
     border-right: 1px solid var(--pb);
     display: flex;
@@ -109,11 +109,12 @@
     text-overflow: ellipsis;
 }
 
-/* ── THREAD ─────────────────────────────────────────────────────── */
+/* THREAD */
 .chat-thread {
     display: flex;
     flex-direction: column;
     overflow: hidden;
+    min-width: 0;
 }
 
 .chat-thread-header {
@@ -185,7 +186,7 @@
 
 .chat-bubble-row.friend .chat-ts { text-align: left; }
 
-/* ── LOAD OLDER ─────────────────────────────────────────────────── */
+/* LOAD OLDER */
 .chat-load-older {
     text-align: center;
     font-size: 11px;
@@ -194,7 +195,7 @@
     flex-shrink: 0;
 }
 
-/* ── EMPTY STATE ────────────────────────────────────────────────── */
+/* EMPTY STATE */
 .chat-empty {
     flex: 1;
     display: flex;
@@ -207,7 +208,7 @@
     text-align: center;
 }
 
-/* ── ERROR BANNER (shown above messages on fetch failure) ───────── */
+/* ERROR BANNER */
 .chat-error {
     margin: 0 0 8px;
     padding: 8px 12px;
@@ -218,9 +219,32 @@
     font-family: var(--fb);
     font-size: 12px;
     line-height: 1.4;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 10px;
 }
 
-/* ── CONNECTION HINT (shown above input while WS is not OPEN) ───── */
+.chat-error-retry {
+    flex-shrink: 0;
+    background: transparent;
+    border: 1px solid rgba(255, 77, 77, 0.5);
+    color: #ff8a80;
+    padding: 3px 10px;
+    border-radius: 4px;
+    font-family: var(--fb);
+    font-size: 11px;
+    font-weight: 600;
+    cursor: pointer;
+    transition: background 0.15s, color 0.15s;
+}
+
+.chat-error-retry:hover {
+    background: rgba(255, 77, 77, 0.18);
+    color: #ffb4b4;
+}
+
+/* CONNECTION HINT */
 .chat-conn-hint {
     padding: 6px 18px;
     border-top: 1px solid var(--pb);
@@ -233,7 +257,7 @@
     flex-shrink: 0;
 }
 
-/* ── INPUT ──────────────────────────────────────────────────────── */
+/* INPUT */
 .chat-input-row {
     display: flex;
     gap: 10px;
