+++ client/src/pages/GamePage/SubviewsGamePage/LobbyChat/LobbyChat.css
@@ -1,3 +1,6 @@
+/* Only the outer shell — sidebar/thread/input styling is shared with
+   HomePage via components/chat/chat.css (the .chat-sidebar / .chat-thread
+   classes rendered by ChatInner). */
 .lobby-chat {
     margin-top: 20px;
     display: grid;
@@ -7,18 +10,4 @@
     background: var(--panel);
     overflow: hidden;
     height: 360px;
-}
-
-.lobby-chat-sidebar {
-    border-right: 1px solid var(--pb);
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-
-.lobby-chat-thread {
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-    min-width: 0;
 }
